fix: javascript block not working in firefox
