+++ src/content/blocksHandler/handlerJavascriptCode.js
@@ -1,0 +1,7 @@
+import { jsContentHandler } from '@/newtab/utils/javascriptBlockUtil';
+
+function javascriptCode({ data }) {
+  return jsContentHandler(...data);
+}
+
+export default javascriptCode;
